all: make tgo opt-in
This change changes the tgo support so that it is opt-in. It introduces
two new flags in the scanner and printer package (ScanTgo and ParseTgo)
to enable the tgo mode.

This change also introudces few changes in the ast. It introduces
a new *ast.Text that corresponds to the previous String BasicLit
contained in the ExprStmt. Also this change tweaks few type names and
introudces a new ast.AttrValue interface, which corresponds to the value
node of an attribute (*ast.Text and *ast.TemplateLiteral).